Remove useless DI for CharacteristicsForumSyncer
